+++ src/blocks/Content.ts
@@ -1,6 +1,7 @@
 import { slateEditor } from '@payloadcms/richtext-slate';
 import { Block } from 'payload/types';
 
+import registry from '../fields/richText/registry';
 import venue from '../fields/richText/venue';
 import { width } from '../fields/width';
 
@@ -14,7 +15,7 @@ const Content: Block = {
       type: 'richText',
       editor: slateEditor({
         admin: {
-          elements: ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'ol', 'ul', venue],
+          elements: ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'ol', 'ul', venue, registry],
         },
       }),
       required: true,