+++ src/fields/richText/registry/Element/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const RegistryLeaf: React.FC<{ attributes: any; element: any; children?: React.ReactNode }> = ({
+  attributes,
+  children,
+}) => <div {...attributes}>{children}</div>;
+
+export default RegistryLeaf;