+++ src/fields/richText/registry/index.ts
@@ -1,0 +1,9 @@
+import Button from './Button';
+import Element from './Element';
+
+export default {
+  name: 'registry',
+  Button,
+  Element,
+  plugins: [],
+};